Service Layer validation 도입으로 중복되는 코드 제거 및 모듈화(only return true || false)
