+++ server/service/validation/index.js
@@ -1,0 +1,7 @@
+const authValidation = require('./auth');
+
+
+
+module.exports = {
+    authValidation
+}
